Add ShoppingCart.js
Map over the items in the cart state and get them to display the proper
information in a CartItem component.

This will display a div for each cart item containing the name, price,
and a picture of the item.

If no cart items are chosen the cart will display a message instead.